experts/yWorkflows/component_creator/components/create_docs/requirements.txt
langchain==0.0.93
fastapi==0.92.0
python-dotenv==1.0.0
openai==0.26.5